Second condition added for the search
By adding the "!== false" condition, it has created a new bug. 

Since all the search parameters are not defined as false but as empty string, the query get all the search parameters (WHERE field LIKE "%%"). With this, if there are data set as NULL in the database, these entries are not returned.